move ListingBadge to components
